changed permissions, added table pagelet
introduced permissions quotationtool.commentary.Read and
quotationtool.commentary.Overview. Read is granted to the
quotationtool.Visitor role etc. So it is easily possible to set up an
app which is closed to anauthenticated users.

Overview is granted to Editors and Managers only.

Added a table pagelet to show all comments in the database. It is only
viewable for principals which have the Overview permission (per
default only Editors and Managers).